feat(slider): support custom label
