+++ packages/elemental-theme/src/custom-elements/ef-slider-marker.less
@@ -1,0 +1,8 @@
+:host {
+  width: @slider-step-width;
+  height: @slider-step-width;
+
+  [part='marker'] {
+    background-color: var(--marker-color, @global-text-color);
+  }
+}
