+++ packages/elements/src/slider/__demo__/index.html
@@ -102,6 +102,33 @@
       <ef-slider range min="0.1" max="10" value="0" step="0.5" show-steps show-input-field></ef-slider>
     </demo-block>
 
+    <demo-block layout="normal" header="Custom label">
+      <ef-slider>
+        <ef-slider-marker value="11"></ef-slider-marker>
+        <ef-slider-marker value="26"></ef-slider-marker>
+        <ef-slider-marker value="67"></ef-slider-marker>
+      </ef-slider>
+      <ef-slider step="5">
+        <ef-slider-marker value="5">5°C</ef-slider-marker>
+        <ef-slider-marker value="20">20°C</ef-slider-marker>
+        <ef-slider-marker value="50">50°C</ef-slider-marker>
+        <ef-slider-marker value="100">100°C</ef-slider-marker>
+      </ef-slider>
+      <ef-slider step="5" range>
+        <ef-slider-marker value="5">5°C</ef-slider-marker>
+        <ef-slider-marker value="20">20°C</ef-slider-marker>
+        <ef-slider-marker value="50">50°C</ef-slider-marker>
+        <ef-slider-marker value="100">100°C</ef-slider-marker>
+      </ef-slider>
+      <ef-slider step="5" range show-input-field>
+        <ef-slider-marker value="0">0°C</ef-slider-marker>
+        <ef-slider-marker value="5">5°C</ef-slider-marker>
+        <ef-slider-marker value="20">20°C</ef-slider-marker>
+        <ef-slider-marker value="50">50°C</ef-slider-marker>
+        <ef-slider-marker value="100">100°C</ef-slider-marker>
+      </ef-slider>
+    </demo-block>
+
     <script>
       const demoBlock = document.querySelectorAll('demo-block');
       demoBlock.forEach(function (block) {
