+++ packages/halo-theme/src/custom-elements/ef-slider-marker.less
@@ -1,0 +1,1 @@
+@import '@refinitiv-ui/elemental-theme/src/custom-elements/ef-slider-marker';
